Add new command register to automatically block duplicate commands
There are issues with commands that use the InputBox. If the inputbox is closed by clicking out of it, the command is technically still running as the function has not ended, but there is no way for the function to continue as it is waiting on an input box that the user can no longer see.